requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
itsdangerous==0.24
virtualenv==1.11.6
wsgiref==0.1.2
Flask-SQLAlchemy==2.0
MySQLdb==1.2.5